Refactor sign-in business logic for DRMs
