+++ Simplified/Reader2/ReaderStackConfiguration/AdobeDRM/AdobeDRMContainer.mm
@@ -42,7 +42,7 @@ static id acsdrm_lock = nil;
     }
 
     // *_rights.xml file contents
-    NSString *rightsPath = [NSString stringWithFormat:@"%@%@", path, RIGHTS_XML_SUFFIX];
+    NSString *rightsPath = [NSString stringWithFormat:@"%@%@", path, ADOBE_RIGHTS_XML_SUFFIX];
     NSData *rightsData = [NSData dataWithContentsOfFile:rightsPath];
     size_t rightsLen = rightsData.length;
     unsigned char *rightsContent = (unsigned char *)rightsData.bytes;
